Addressed PR comments #1
